Doc edits
Added some edits. Use "can" or "might" rather than "may" because "may" generally indicates permission. I added in a few commas. Finally, if you use the words "this, that, those, or these" you need to have a noun after them. Otherwise, they are hard to translate and also can lead to confusion as to what "this" is referring to.

Also, I found this para: 
The CBQRI registers are defined so that the software can perform two
individual 4 byte accesses, or the hardware can perform two independent 4 byte
transactions, resulting from an 8 byte access, to the high and low halves of the
register as long as the register semantics, with regards to side-effects, are
respected between the two software accesses, or two hardware transactions,
respectively.

to be confusing until I took out the two clauses. If those clauses are needed, let me know and I'll take another pass.

Signed-off-by: Kersten Richter <kersten@riscv.org>